fix: linker error in sysinfo
