+++ src/data_provider/src/sysInfoWin.cpp
@@ -974,181 +974,3 @@ nlohmann::json SysInfo::getHotfixes() const
 
     return ret;
 }
-
-// This function provides a minimal implementation for converting C strings to BSTRs,
-// avoiding the need to include a full COM library. This can be useful when you only need this specific functionality
-// and want to minimize dependencies.
-namespace _com_util
-{
-    BSTR ConvertStringToBSTR(const char* str)
-    {
-        int len = MultiByteToWideChar(CP_ACP, 0, str, -1, NULL, 0);
-        BSTR bstr = SysAllocStringLen(0, len);
-        MultiByteToWideChar(CP_ACP, 0, str, -1, bstr, len);
-        return bstr;
-    }
-}
-// The BstrToString function takes a Windows-specific string (BSTR) and converts it into
-// a standard C++ string (std::string) that can be more easily used in the application.
-std::string BstrToString(BSTR bstr)
-{
-    std::wstring_convert<std::codecvt_utf8<wchar_t>> converter;
-    std::wstring wstr(bstr, SysStringLen(bstr));
-    return converter.to_bytes(wstr);
-}
-
-void QueryWMIHotFixes(std::set<std::string>& hotfixSet, IComHelper& comHelper)
-{
-    HRESULT hres;
-    IWbemLocator* pLoc = NULL;
-    IWbemServices* pSvc = NULL;
-    std::ostringstream oss;
-
-    hres = comHelper.CreateWmiLocator(pLoc);
-
-    if (FAILED(hres))
-    {
-        oss << "WMI: Error creating IWbemLocator. Code: " << std::hex << hres;
-        throw std::runtime_error(oss.str());
-    }
-
-    hres = comHelper.ConnectToWmiServer(pLoc, pSvc);
-
-    if (FAILED(hres))
-    {
-        if (pLoc) pLoc->Release();
-
-        oss << "WMI: connection failed. Code: " << std::hex << hres;
-        throw std::runtime_error(oss.str());
-    }
-
-    hres = comHelper.SetProxyBlanket(pSvc);
-
-    if (FAILED(hres))
-    {
-        if (pSvc) pSvc->Release();
-
-        if (pLoc) pLoc->Release();
-
-        oss << "WMI: security error. Code: " << std::hex << hres;
-        throw std::runtime_error(oss.str());
-    }
-
-    IEnumWbemClassObject* pEnumerator = NULL;
-    hres = comHelper.ExecuteWmiQuery(pSvc, pEnumerator);
-
-    if (FAILED(hres))
-    {
-        if (pLoc) pLoc->Release();
-
-        if (pLoc) pSvc->Release();
-
-        oss << "WMI: query error. Code: " << std::hex << hres;
-        throw std::runtime_error(oss.str());
-    }
-
-    IWbemClassObject* pclsObj = NULL;
-    ULONG uReturn = 0;
-
-    while (pEnumerator)
-    {
-        HRESULT hr = pEnumerator->Next(WBEM_INFINITE, 1, &pclsObj, &uReturn);
-
-        if (0 == uReturn)
-        {
-            break;
-        }
-
-        VARIANT vtProp;
-        hr = pclsObj->Get(L"HotFixID", 0, &vtProp, 0, 0);
-
-        if (SUCCEEDED(hr) && vtProp.vt == VT_BSTR)
-        {
-            std::string hotfix = BstrToString(vtProp.bstrVal);
-
-            if (hotfixSet.find(hotfix) == hotfixSet.end())
-            {
-                // New HotFix found
-                hotfixSet.insert(hotfix);
-            }
-        }
-
-        VariantClear(&vtProp);
-        pclsObj->Release();
-    }
-
-    pSvc->Release();
-    pLoc->Release();
-    pEnumerator->Release();
-}
-
-void QueryWUHotFixes(std::set<std::string>& hotfixSet, IComHelper& comHelper)
-{
-    HRESULT hres;
-    IUpdateSearcher* pUpdateSearcher = NULL;
-    IUpdateHistoryEntryCollection* pHistory = NULL;
-    std::regex hotfixRegex("KB[0-9]+");
-    std::ostringstream oss;
-
-    hres = comHelper.CreateUpdateSearcher(pUpdateSearcher);
-
-    if (FAILED(hres))
-    {
-        oss << "WUA: UpdateSearcher error. Code: " << std::hex << hres;
-        throw std::runtime_error(oss.str());
-    }
-
-    LONG count;
-    hres = comHelper.GetTotalHistoryCount(pUpdateSearcher, count);
-
-    if (FAILED(hres))
-    {
-        if (pUpdateSearcher) pUpdateSearcher->Release();
-
-        oss << "WUA: Error getting total update history count. Code: 0x" << std::hex << hres;
-        throw std::runtime_error(oss.str());
-    }
-
-    hres = comHelper.QueryHistory(pUpdateSearcher, pHistory, count);
-
-    if (FAILED(hres))
-    {
-        if (pUpdateSearcher) pUpdateSearcher->Release();
-
-        oss << "WUA: Error querying update history. Code: 0x" << std::hex << hres;
-        throw std::runtime_error(oss.str());
-    }
-
-    LONG historyCount;
-    hres = comHelper.GetCount(pHistory, historyCount);
-
-    for (LONG i = 0; i < historyCount; ++i)
-    {
-        IUpdateHistoryEntry* pEntry = NULL;
-        comHelper.GetItem(pHistory, i, &pEntry);
-        BSTR title;
-        comHelper.GetTitle(pEntry, title);
-        std::string titleStr = BstrToString(title);
-
-        std::smatch match;
-
-        if (std::regex_search(titleStr, match, hotfixRegex))
-        {
-            std::string hotfix = match[0];
-
-            if (hotfixSet.find(hotfix) == hotfixSet.end())
-            {
-                // New HotFix found
-                hotfixSet.insert(hotfix);
-            }
-        }
-
-        if (pEntry) pEntry->Release();
-
-        if (SysFreeString) SysFreeString(title);
-    }
-
-    if (pHistory) pHistory->Release();
-
-    if (pUpdateSearcher) pUpdateSearcher->Release();
-}
